+++ reana_server/rest/workflows.py
@@ -59,6 +59,7 @@ blueprint = Blueprint("workflows", __name__)
         "size": fields.Int(validate=validate.Range(min=1)),
         "include_progress": fields.Bool(location="query"),
         "include_workspace_size": fields.Bool(location="query"),
+        "include_retention_rules": fields.Bool(location="query"),
         "workflow_id_or_name": fields.Str(),
     }
 )
@@ -124,6 +125,10 @@ def get_workflows(user, **kwargs):  # noqa
         - name: include_workspace_size
           in: query
           description: Include size information of the workspace.
+          type: boolean
+        - name: include_retention_rules
+          in: query
+          description: Include workspace retention rules of the workflows.
           type: boolean
         - name: workflow_id_or_name
           in: query